pyproject.toml
[project]
name = "medusa"
version = "0.1.0"
description = "Minimal Python static site generator"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "click>=8.1",
    "jinja2>=3.1",
    "mistune>=3.0",
    "pyyaml>=6.0",
    "watchdog>=4.0",
    "websockets>=12.0",
    "rjsmin>=1.2",
    "pillow>=10.0",
]

[project.scripts]
medusa = "medusa.cli:main"

[tool.setuptools]
packages = ["medusa"]

[build-system]
requires = ["setuptools>=67", "wheel"]
build-backend = "setuptools.build_meta"
